Implemented style fixes
