+++ src/compilation/build_module.cc
@@ -1,0 +1,253 @@
+/*!
+ *  Copyright (c) 2017 by Contributors
+ *  Compile executable modules.
+ * \file build_module.cc
+ */
+#include "./tvm/build_module.h"
+#include "./tvm/operation.h"
+#include "./tvm/ir_pass.h"
+#include "./tvm/codegen.h"
+
+
+namespace tvm {
+
+std::string Target::str() const {
+    std::ostringstream result;
+    result << target_name;
+    for (const auto &x : options) {
+        result << " " << x;
+    }
+    return result.str();
+}
+
+
+namespace target {
+Target llvm() {
+    std::unordered_set<std::string> keys({ "llvm", "cpu" });
+    std::unordered_set<std::string> options;
+    return Target("llvm", kDLCPU, 512, 1, keys, options);
+}
+
+Target cuda() {
+    std::unordered_set<std::string> keys({ "cuda", "gpu" });
+    std::unordered_set<std::string> options;
+    return Target("cuda", kDLGPU, 512, 32, keys, options);
+}
+
+Target rocm() {
+    std::unordered_set<std::string> keys({ "rocm", "gpu" });
+    std::unordered_set<std::string> options;
+    return Target("rocm", kDLROCM, 256, 1, keys, options);
+}
+
+Target metal() {
+    std::unordered_set<std::string> keys({ "gpu" });
+    std::unordered_set<std::string> options;
+    return Target("metal", kDLMetal, 256, 1, keys, options);
+}
+
+Target rasp() {
+    std::unordered_set<std::string> keys({ "llvm", "cpu" });
+    std::unordered_set<std::string> options({
+        "-device=rasp",
+        "-mtriple=armv7l-none-linux-gnueabihf",
+        "-mcpu=cortex-a53",
+        "-mattr=+neon"
+    });
+    return Target("llvm", kDLCPU, 512, 1, keys, options);
+}
+
+Target stackvm() {
+    std::unordered_set<std::string> keys({ "stackvm", "cpu" });
+    std::unordered_set<std::string> options;
+    return Target("stackvm", kDLCPU, 512, 1, keys, options);
+}
+}  // namespace target
+
+
+
+bool LLVMEnabled() {
+    const runtime::PackedFunc* pf = runtime::Registry::Get("codegen.llvm_target_enabled");
+    if (pf == nullptr) return false;
+    return (*pf)("llvm");
+}
+
+/*! \return The default host target for a given device target */
+Target DefaultTargetHost(Target target) {
+    if (target.device_type == kDLCPU) {
+        return target;
+    } else {
+        if (LLVMEnabled()) {
+            return target::llvm();
+        } else {
+            return target::stackvm();
+        }
+    }
+}
+
+
+Buffer BufferWithOffsetAlignment(Array<Expr> shape, Type dtype, std::string name,
+    int data_alignment, int offset_factor) {
+    auto data = Var(name, Handle());
+
+    Expr elem_offset;
+    if (offset_factor != 0) {
+        elem_offset = Var(name + "_elem_offset", shape[0].type());
+    } else {
+        elem_offset = Expr();
+    }
+
+    return BufferNode::make(data, dtype, shape, Array<Expr>(), elem_offset, name, "",
+        data_alignment, offset_factor);
+}
+
+void GetBinds(const Array<Tensor>& args, const std::unordered_map<Tensor, Buffer>& binds,
+    Map<Tensor, Buffer>* out_binds, Array<NodeRef>* out_arg_list,
+    const BuildConfig& config) {
+    *out_binds = binds;
+
+    for (const auto &x : args) {
+        if (out_binds->find(x) == out_binds->end()) {
+            auto buf = BufferWithOffsetAlignment(x->shape, x->dtype, x->op->name,
+                config.data_alignment, config.offset_factor);
+            out_binds->Set(x, buf);
+            out_arg_list->push_back(buf);
+        } else {
+            out_arg_list->push_back((*out_binds)[x]);
+        }
+    }
+}
+
+
+/*!
+* \brief Build a Stmt given a schedule, args and binds. This function runs the IR passes.
+* \param sch The schedule to build.
+* \param args The arguments for the schedule.
+* \param binds Buffer assignments.
+* \param loopPartition True if the LoopPartition pass should be included.
+* \param argListOut Returns the arguments for the Stmt.
+* \param config The build configuration.
+* \return The built Stmt.
+*/
+Stmt BuildStmt(Schedule sch, const Array<Tensor>& args,
+    const std::unordered_map<Tensor, Buffer>& binds, bool loop_partition,
+    Array<NodeRef> *out_arg_list, const BuildConfig& config) {
+    Map<Tensor, Buffer> out_binds;
+    GetBinds(args, binds, &out_binds, out_arg_list, config);
+
+    sch = sch.normalize();
+
+    // Phase 0
+
+    auto bounds = schedule::InferBound(sch);
+    auto stmt = schedule::ScheduleOps(sch, bounds);
+    stmt = ir::InjectPrefetch(stmt);
+
+
+    // Phase 1
+
+    stmt = ir::StorageFlatten(stmt, out_binds, 64);
+    stmt = ir::CanonicalSimplify(stmt);
+    if (loop_partition) {
+        stmt = ir::LoopPartition(stmt);
+    }
+    stmt = ir::VectorizeLoop(stmt);
+    stmt = ir::InjectVirtualThread(stmt);
+    stmt = ir::InjectDoubleBuffer(stmt, config.double_buffer_split_loop);
+    stmt = ir::StorageRewrite(stmt);
+    stmt = ir::UnrollLoop(stmt, config.auto_unroll_max_step, config.auto_unroll_max_depth,
+        config.auto_unroll_max_extent, config.unroll_explicit);
+
+
+    // Phase 2
+
+    stmt = ir::Simplify(stmt);
+    stmt = ir::LowerStorageAccessInfo(stmt);
+    stmt = ir::RemoveNoOp(stmt);
+    stmt = ir::RewriteUnsafeSelect(stmt);
+
+    return stmt;
+}
+
+LoweredFunc lower(Schedule sch, const Array<Tensor>& args, const std::string& name,
+    const std::unordered_map<Tensor, Buffer>& binds, const BuildConfig& config) {
+    Array<NodeRef> out_arg_list;
+    auto stmt = BuildStmt(sch, args, binds, true, &out_arg_list, config);
+    return ir::MakeAPI(stmt, name, out_arg_list, 0, config.restricted_func);
+}
+
+runtime::Module build(const Array<LoweredFunc>& funcs, const Target& target,
+    Target* target_host, const BuildConfig& config) {
+    std::unordered_set<std::string> all_names;
+    for (const auto &x : funcs) {
+        CHECK(all_names.count(x->name) == 0) << "Duplicate function name " << x->name;
+        all_names.insert(x->name);
+    }
+
+    Target target_host_val = target_host == nullptr ?
+        DefaultTargetHost(target) :
+        *target_host;
+
+    Array<LoweredFunc> fhost;
+    Array<LoweredFunc> fdevice;
+
+    for (const auto &x : funcs) {
+        if (x->func_type == kMixedFunc) {
+            auto func = x;
+            if (config.detect_global_barrier) {
+                func = ir::ThreadSync(func, "global");
+            }
+
+            func = ir::ThreadSync(func, "shared");
+            func = ir::LowerThreadAllreduce(func, target.thread_warp_size);
+            auto fsplits = ir::SplitHostDevice(func);
+            fhost.push_back(fsplits[0]);
+            for (auto f = fsplits.begin() + 1; f != fsplits.end(); ++f) {
+                fdevice.push_back(*f);
+            }
+        } else if (x->func_type == kHostFunc) {
+            fhost.push_back(x);
+        } else if (x->func_type == kDeviceFunc) {
+            fdevice.push_back(x);
+        } else {
+            CHECK(false) << "unknown function type " << x->func_type;
+        }
+    }
+
+    if (target.keys.count("gpu") > 0 && fdevice.size() == 0) {
+        LOG(WARNING) << "Specified target " + target.str() +
+            " but cannot find device code. Did you forget to bind?";
+    }
+
+    for (int i = 0; i < fhost.size(); ++i) {
+        auto func = fhost[i];
+        func = ir::BindDeviceType(func, target.device_type);
+        func = ir::LowerTVMBuiltin(func);
+        fhost.Set(i, func);
+    }
+
+
+    for (int i = 0; i < fdevice.size(); ++i) {
+        auto func = fdevice[i];
+        func = ir::LowerIntrin(func, target.target_name);
+        fdevice.Set(i, func);
+    }
+
+    for (int i = 0; i < fhost.size(); ++i) {
+        auto func = fhost[i];
+        func = ir::LowerIntrin(func, target_host_val.target_name);
+        func = ir::CombineContextCall(func);
+        fhost.Set(i, func);
+    }
+
+    auto mhost = codegen::Build(fhost, target_host_val.str());
+
+    if (fdevice.size() > 0) {
+        auto mdev = codegen::Build(fdevice, target.str());
+        mhost.Import(mdev);
+    }
+
+    return mhost;
+}
+
+}  // namespace tvm
